Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,7 +2,7 @@
 int main()
 {
     int a,b,c;
-    int result= (a+b+c)/3;
+    int result= (float)(a+b+c)/3;
     scanf("%d %d %d", &a,&b,&c);
-    printf("Average: %.2f", (float)result);
+    printf("Average: %.2f", result);
 }
